1.5.1 ocpc: 新增sd卡回看进度实现，优化逻辑

+++ AgoraIotLink/Source/Utils/Extensions/Extension+Fundation.swift
@@ -62,9 +62,9 @@ public extension String {
         return timestamp
     }
     
-    static func dateTimeSpace(_ lastTime : Int)->Int {
+    static func dateTimeSpace(_ lastTime : UInt64)->UInt64 {
         let timeInterval: TimeInterval = Date().timeIntervalSince1970
-        let timestamp = Int(timeInterval)
+        let timestamp = UInt64(timeInterval)
         log.i("dateTimeSpace currentTime:\(timestamp)")
         return timestamp - lastTime
     }
